Unravel/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-cors-headers==3.10.0
django-extensions==3.1.5
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.3
oauthlib==3.1.1
pandas==1.4.2
Pillow==8.4.0
plotly==5.8.0
pycparser==2.20
pydot==1.4.2
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tenacity==8.0.1
uritemplate==4.1.1
urllib3==1.26.7
